requirements.txt
langgraph
langchain
langchain_google_genai
pandas
numpy
scikit-learn
sdv
python-dotenv
google-genai